@@ -1,16 +1,19 @@
 {
   "name": "liquorsunlimited",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
-    "eslint-config-rallycoding": "^3.2.0",
-    "google-map-react": "^0.29.0",
-    "react": "^16.2.0",
-    "react-bootstrap": "^0.32.0",
-    "react-dom": "^16.2.0",
-    "react-google-maps": "^9.4.5",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.0"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@emailjs/browser": "^4.4.1",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@react-google-maps/api": "^2.19.3",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.4",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.1",
+    "react-scripts": "5.0.1",
+    "sass": "^1.77.8"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -18,5 +21,20 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "devDependencies": {}
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ],
+    "sideEffects": [
+    "*.css", 
+    "*.scss"
+  ]
+  }
 }
